make mingw32 compat
by careful use of uintmax_t which is 64bit even on 32bit mingw

specifically the offset_t needs to be consistently 64bit, because it
becomes serialized in the index

myabe consider actually making it std::uint64_t in future

however, the mmap cannot be accessed with 64bit values on mingw32, so
we need to downcast each access. Some runtime checks might be
appropriate here in future

+++ include/binfuse/filter.hpp
@@ -249,7 +249,7 @@ private:
   }
 
   // returns existing file size
-  std::size_t ensure_file()
+  std::uintmax_t ensure_file()
     requires(AccessMode == mio::access_mode::write)
   {
     if (filepath_.empty()) {
@@ -257,7 +257,7 @@ private:
                                "'");
     }
 
-    std::size_t existing_filesize = 0;
+    std::uintmax_t existing_filesize = 0;
     if (std::filesystem::exists(filepath_)) {
       existing_filesize = std::filesystem::file_size(filepath_);
     } else {
